requirements.txt
py-clob-client==0.17.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.3
python-dateutil==2.9.0
colorlog==6.8.2
